refactor config idp buttons
